Refactor: move profile models to profile/models package
