+++ app/src/main/java/fr/free/nrw/commons/mwapi/OkHttpJsonApiClient.java
@@ -13,10 +13,10 @@ import fr.free.nrw.commons.location.LatLng;
 import fr.free.nrw.commons.nearby.Place;
 import fr.free.nrw.commons.nearby.model.NearbyResponse;
 import fr.free.nrw.commons.nearby.model.NearbyResultItem;
-import fr.free.nrw.commons.data.models.profile.FeaturedImages;
+import fr.free.nrw.commons.profile.models.FeaturedImages;
 import fr.free.nrw.commons.profile.achievements.FeedbackResponse;
-import fr.free.nrw.commons.data.models.profile.LeaderboardResponse;
-import fr.free.nrw.commons.data.models.profile.UpdateAvatarResponse;
+import fr.free.nrw.commons.profile.models.LeaderboardResponse;
+import fr.free.nrw.commons.profile.models.UpdateAvatarResponse;
 import fr.free.nrw.commons.upload.FileUtils;
 import fr.free.nrw.commons.upload.structure.depictions.DepictedItem;
 import fr.free.nrw.commons.utils.ConfigUtils;
