+++ app/src/main/java/fr/free/nrw/commons/profile/models/Achievements.kt
@@ -1,4 +1,4 @@
-package fr.free.nrw.commons.data.models.profile
+package fr.free.nrw.commons.profile.models
 
 import fr.free.nrw.commons.profile.achievements.FeedbackResponse
